Move Mac DNS to own namespace
