+++ src/Installer/DNS/Mac/DnsDock.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Dock\Installer\DNS;
+namespace Dock\Installer\DNS\Mac;
 
 use Dock\Dinghy\SshClient;
 use Dock\Installer\InstallContext;
